First Iteration of Items
There are issues, with getting the prefab able to refer to the stamina
bar. I want to implement it without having to search for it. I will need
to figure out a soltuion to that. I as well have to figure out how to
store/items into invetory slots. Finally Need to figure out how ot pick
up item and delete and drop item (respawn into the world).

First figure out how to create prefab for item and have it refer to the
stamina bar